macOS fixes

+++ recipe/build.sh
@@ -219,8 +219,10 @@ fi
 if [[ ${HOST} =~ .*darwin.* ]]; then
 
     # Avoid Xcode
-    cp "${RECIPE_DIR}"/xcrun .
-    cp "${RECIPE_DIR}"/xcodebuild .
+    [[ -f xcrun ]] || cp "${RECIPE_DIR}"/xcrun .
+    [[ -f xcodebuild ]] || cp "${RECIPE_DIR}"/xcodebuild .
+    [[ -f xcode-select ]] || cp "${RECIPE_DIR}"/xcode-select .
+    chmod +x xcrun xcodebuild xcode-select
     # Some test runs 'clang -v', but I do not want to add it as a requirement just for that.
     ln -s "${CXX}" ${HOST}-clang || true
     # For ltcg we cannot use libtool (or at least not the macOS 10.9 system one) due to lack of LLVM bitcode support.
